--- src/Agent.Service/Linux/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "vstsagent-service",
-  "description": "VSTS Agent Service module to run in systemd",
-  "version": "0.0.1",
-  "preferGlobal": true,
-  "license": "MIT",
-  "dependencies": {
-      "shelljs": "0.3.x",
-      "moment": "2.12.x"
-  },
-  "devDependencies": {
-    "del": "^1.1.1",
-    "minimist": "0.2.0",
-    "gulp": "^3.8.11",
-    "gulp-gzip": "^1.0.0",
-    "gulp-mocha": "^2.0.0",
-    "gulp-tar": "^1.4.0",
-    "gulp-tsc": "^0.10.1",
-    "merge2": "^0.3.1",
-    "mocha": "^2.1.0"
-  }
-}
