+++ src/Misc/layoutbin/vsts.agent.service.template
@@ -4,7 +4,6 @@ After=network.target
 
 [Service]
 ExecStart=/usr/bin/node {BinDirectory}/AgentService.js
-Restart=always
 User={User}
 Environment=PATH=/user/bin:/usr/local/bin
 Environment=NODE_ENV=production
